refactor: refactor calendar
